refactor: board 도메인 타입을 공통 타입으로 추출
